% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{patch_landscapes_from_image}
\alias{patch_landscapes_from_image}
\title{Sample patches from an image and compute landscapes from them}
\usage{
patch_landscapes_from_image(
  image_name,
  number_of_patches,
  pixel_radius_for_patches = 50,
  number_of_cores = 1,
  patch_center_image = FALSE,
  return_patches = FALSE,
  remove_patch_data = TRUE,
  return_samples = FALSE,
  lower_threshold = 0,
  upper_threshold = 1,
  proportion_of_patch_sparse = 0.025,
  noise_threshold = 0.05,
  number_of_layers = 50
)
}
\arguments{
\item{image_name}{File path (absolute or relative to current directory) to the image. Supported file formats are same as OpenImageR: .png, .jpeg
.jpg, and .tiff}

\item{number_of_patches}{Number of patches to sample from image}

\item{pixel_radius_for_patches}{All sampled patches will be of this radius, default is 50 pixels.}

\item{number_of_cores}{The number of CPU cores to parallelize over when computing landscapes for this image. Default is 1 (no parallelization).}

\item{patch_center_image}{File path (absolute or relative to current directory) to an image to use for selecting the centers of patches. Useful
if you have a masked version of the image where you want patches to be centered. Defaults to FALSE, in which case the image at image_name is
used for patch centers.}

\item{return_patches}{If set to TRUE, output is a list image_landscape_results where image_landscape_results$data is the same data matrix as the default output,
image_landscape_results$patches is a list of patch objects (in the same order as the rows of the data matrix) that describe the center and position in the
original image for each patch. Default is FALSE, in which case output is described in Value below.}

\item{remove_patch_data}{If set to FALSE and return_patches is TRUE, the patch details also include the raw subset of the image comprising each patch.
Useful for some imaging tasks. WARNING: Memory costs can quickly get out of hand when setting this to TRUE. Default is FALSE, in which case such data
is not retained.}

\item{return_samples}{If set to TRUE and return_patches is TRUE, will also append the samples to each patch entry. Default is FALSE.
Useful for some imaging tasks. WARNING: Memory costs can quickly get out of hand when setting this to TRUE. Default is FALSE, in which case such data
is not retained.}

\item{noise_threshold}{A number between 0 or 1 representing a percentage. The bottom noise_threshold pixels in the image by intensity will be set to 0.
Used to eliminate small intensity background noise. Default is .05, i.e. 5\%.}

\item{number_of_layers}{For TDA experts: the maximum landscape Lambda_{number_of_layers} to compute. Default is 50.}
}
\value{
Returns a sparse matrix image_landscape_results with number_of_patches rows. image_landscape_results\[i,\] is the landscape vector computed for the
i'th sampled patch.

Several base R matrix operations are not implemented for SparseM matrices. In this case you may need to use as.matrix(image_landscape_results) instead.
Be warned, however, that the memory consumption will generally be much larger.

If the matrix has too many entries for SparseM, then instead a regular matrix is output.
}
\description{
The main pipeline function of TDAExplore for programmatic access. Most users will only be interested in the first 4 parameters.
A few parameters are undocumented and it is inadvisable to change them.
}
\examples{
computed_landscapes <- patch_landscapes_from_image("test/triangles.jpg",10)
# Plot the average landscape vector
plot(colMeans(as.matrix(computed_landscapes)))

computed_landscapes <- patch_landscapes_from_image("test/triangles.jpg",10,return_patches=TRUE)
# Where was the center of the 5th patch?
computed_landscapes$patches[[5]]$center
}
